+++ src/app/features/sections/main-section/purchases/purchase-card/purchase-card-fields/purchase-card-fields.component.scss
@@ -1,3 +1,5 @@
+@use '../../../../../../../../public/theme/boeris-theme';
+
 .status {
 
 }
@@ -5,36 +7,61 @@
 .fields {
     display: flex;
     flex-direction: row;
+    justify-content: space-evenly;
+    height: 100%;
+    align-items: baseline;
 }
 
 .field-icon {
     display: flex;
     flex-direction: row;
-    justify-content: space-evenly;
+    align-items: center;
+}
+
+i {
+    margin-right: 5px;
+}
+
+p {
+    margin: 0;
+    user-select: none;
 }
 
 .left-panel {
     display: flex;
     flex-direction: column;
+    width: 40%;
 }
 
 .right-panel {
     display: flex;
-    flex-direction: row;
+    flex-direction: column;
+    width: 60%;
+    height: 100%;
 }
 
-.vertical-fields {
+.row-fields {
     display: flex;
-    flex-direction: column;
+    flex-direction: row;
     justify-content: space-between;
 }
 
-.empty {
+.partner-price {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    height: 10%;
+}
+
+.dates {
     display: flex;
     flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    height: 100%;
+    justify-content: start;
+    height: 90%;
+}
+
+.empty {
+    height: 15px;
 }
 
 .id-provider {
@@ -46,5 +73,31 @@
 .description {
     display: flex;
     flex-direction: column;
-    width: 50%;
+}
+
+.text-description-overflowed {
+    transition: background-color 0.3s ease;
+    cursor: pointer;
+}
+
+.text-description-overflowed:hover {
+    background-color: boeris-theme.$secondaryDarkColor;
+}
+
+.text-description-overflowed:active {
+    background-color: boeris-theme.$secondaryDarkerColor;
+}
+
+::ng-deep {
+    .p-card.p-component {
+        height: 135px;
+    }
+
+    .p-card-body {
+        height: 135px;
+    }
+
+    .p-card-content {
+        height: calc(135px - 40px);
+    }
 }
