+++ src/app/features/sections/main-section/purchases/purchase-card/purchase-card.component.html
@@ -1,8 +1,10 @@
-<p-card>
-    <app-purchase-card-fields
-    [purchase]="purchase"
-    ></app-purchase-card-fields>
-    <div class="buttons">
-        
+<p-card styleClass="custom-card">
+    <div class="content">
+        <app-purchase-card-fields
+        [purchase]="purchase()"
+        ></app-purchase-card-fields>
+        <div class="buttons">
+    
+        </div>
     </div>
 </p-card>
