+++ src/app/features/sections/main-section/purchases/purchase-card/purchase-card.component.scss
@@ -1,0 +1,11 @@
+.custom-card {
+    overflow: hidden;
+    position: relative;
+}
+
+.content {
+    width: 100%;
+    height: 100%;
+
+    overflow: hidden;
+}
