+++ src/app/features/sections/main-section/purchases/purchases.component.html
@@ -1,1 +1,7 @@
-<p>purchases works!</p>
+<!-- <app-purchase-card
+[purchase]="purchase1"
+></app-purchase-card>
+<div class="empty"><p>Hola</p></div> -->
+<app-purchase-card
+[purchase]="purchase2"
+></app-purchase-card>
