+++ src/app/features/sections/main-section/purchases/purchases.component.scss
@@ -1,0 +1,3 @@
+:host {
+    width: 100%;
+}
